requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
aiofiles>=23.0.0
PyMuPDF>=1.20.0
PyPDF2>=3.0.0
python-docx>=1.0.0
